+++ app/src/main/java/com/antoniogabriel/healthylife/RcqinputActivity.kt
@@ -16,9 +16,9 @@ class RcqinputActivity : AppCompatActivity() {
 
         btnResultRcq.setOnClickListener {
             if(!isEmpty(edtCintura) && !isEmpty(edtQuadril)){
-                val cintura = edtCintura.text.toString().toDouble()
-                val quadril = edtQuadril.text.toString().toDouble()
-                val resultadoRcq = String.format("%.2f", cintura/quadril).toDouble()
+                val cintura = edtCintura.text.toString().toFloat()
+                val quadril = edtQuadril.text.toString().toFloat()
+                val resultadoRcq = String.format("%.2f", cintura/quadril)
                 val transportResultRcq = "$resultadoRcq"
                 startActivity(Intent(this@RcqinputActivity, RcqresultActivity::class.java).apply{
                     putExtra("resultadoRcq", transportResultRcq)
